+++ src/Make_vms.mms
@@ -311,12 +311,13 @@ SRC =	arabic.c autocmd.c beval.c blob.c blowfish.c buffer.c change.c charset.c \
 	crypt.c crypt_zip.c debugger.c dict.c diff.c digraph.c edit.c eval.c \
 	evalfunc.c ex_cmds.c ex_cmds2.c ex_docmd.c ex_eval.c ex_getln.c \
 	if_cscope.c if_xcmdsrv.c fileio.c findfile.c fold.c getchar.c \
-	hardcopy.c hashtab.c indent.c insexpand.c json.c list.c main.c mark.c \
-	menu.c mbyte.c memfile.c memline.c message.c misc1.c misc2.c move.c \
-	normal.c ops.c option.c popupmnu.c popupwin.c profiler.c quickfix.c \
-	regexp.c search.c sha256.c sign.c spell.c spellfile.c syntax.c tag.c \
-	term.c termlib.c testing.c textprop.c ui.c undo.c usercmd.c userfunc.c \
-	version.c screen.c window.c os_unix.c os_vms.c pathdef.c \
+	hardcopy.c hashtab.c highlight.c indent.c insexpand.c json.c list.c \
+	main.c mark.c menu.c mbyte.c memfile.c memline.c message.c misc1.c \
+	misc2.c move.c normal.c ops.c option.c popupmnu.c popupwin.c \
+	profiler.c quickfix.c regexp.c search.c sha256.c sign.c spell.c \
+	spellfile.c syntax.c tag.c term.c termlib.c testing.c textprop.c ui.c \
+	undo.c usercmd.c userfunc.c version.c screen.c window.c os_unix.c \
+	os_vms.c pathdef.c \
 	$(GUI_SRC) $(PERL_SRC) $(PYTHON_SRC) $(TCL_SRC) \
  	$(RUBY_SRC) $(HANGULIN_SRC) $(MZSCH_SRC) $(XDIFF_SRC)
 
@@ -325,14 +326,14 @@ OBJ = 	arabic.obj autocmd.obj beval.obj blob.obj blowfish.obj buffer.obj change.
 	digraph.obj edit.obj eval.obj evalfunc.obj ex_cmds.obj ex_cmds2.obj \
 	ex_docmd.obj ex_eval.obj ex_getln.obj if_cscope.obj if_xcmdsrv.obj \
 	fileio.obj findfile.obj fold.obj getchar.obj hardcopy.obj hashtab.obj \
-	indent.obj insexpand.obj json.obj list.obj main.obj mark.obj \
-	menu.obj memfile.obj memline.obj message.obj misc1.obj misc2.obj \
-	move.obj mbyte.obj normal.obj ops.obj option.obj popupmnu.obj \
-	popupwin.obj profiler.obj quickfix.obj regexp.obj search.obj \
-	sha256.obj sign.obj spell.obj spellfile.obj syntax.obj tag.obj \
-	term.obj termlib.obj testing.obj textprop.obj ui.obj undo.obj usercmd.obj \
-	userfunc.obj screen.obj version.obj window.obj os_unix.obj os_vms.obj \
-	pathdef.obj if_mzsch.obj \
+	highlight.obj indent.obj insexpand.obj json.obj list.obj main.obj \
+	mark.obj menu.obj memfile.obj memline.obj message.obj misc1.obj \
+	misc2.obj move.obj mbyte.obj normal.obj ops.obj option.obj \
+	popupmnu.obj popupwin.obj profiler.obj quickfix.obj regexp.obj \
+	search.obj sha256.obj sign.obj spell.obj spellfile.obj syntax.obj \
+	tag.obj term.obj termlib.obj testing.obj textprop.obj ui.obj undo.obj \
+	usercmd.obj userfunc.obj screen.obj version.obj window.obj \
+	os_unix.obj os_vms.obj pathdef.obj if_mzsch.obj \
 	$(GUI_OBJ) $(PERL_OBJ) $(PYTHON_OBJ) $(TCL_OBJ) \
  	$(RUBY_OBJ) $(HANGULIN_OBJ) $(MZSCH_OBJ) $(XDIFF_OBJ)
 
@@ -596,6 +597,10 @@ hardcopy.obj : hardcopy.c vim.h [.auto]config.h feature.h os_unix.h \
  gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
  globals.h
 hashtab.obj : hashtab.c vim.h [.auto]config.h feature.h os_unix.h \
+ ascii.h keymap.h term.h macros.h structs.h regexp.h \
+ gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
+ globals.h
+highlight.obj : highlight.c vim.h [.auto]config.h feature.h os_unix.h \
  ascii.h keymap.h term.h macros.h structs.h regexp.h \
  gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
  globals.h
