+++ src/globals.h
@@ -355,6 +355,16 @@ EXTERN char_u	hash_removed;
 EXTERN int	scroll_region INIT(= FALSE); /* term supports scroll region */
 EXTERN int	t_colors INIT(= 0);	    /* int value of T_CCO */
 
+// highlight groups for 'highlight' option
+EXTERN garray_T highlight_ga INIT(= {0 COMMA 0 COMMA sizeof(hl_group_T) COMMA 10 COMMA NULL});
+
+#ifdef FEAT_CMDL_COMPL
+// Flags to indicate an additional string for highlight name completion.
+EXTERN int include_none INIT(= 0);	// when 1 include "None"
+EXTERN int include_default INIT(= 0);	// when 1 include "default"
+EXTERN int include_link INIT(= 0);	// when 2 include "link" and "clear"
+#endif
+
 /*
  * When highlight_match is TRUE, highlight a match, starting at the cursor
  * position.  Search_match_lines is the number of lines after the match (0 for
